pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
name = "grid_hooks"
version = "0.0.0"
description = "Pre-commit hooks for Grid."
authors = ["Bilbottom"]
readme = "README.md"
packages = [{include = "grid_hooks"}]

[tool.poetry.scripts]
grid-version-badge = "grid_hooks.grid_version_badge:main"

[tool.poetry.dependencies]
python = "^3.9"
pyyaml = "^6.0.1"
pybadges = "^3.0.1"

[tool.poetry.group]
dev.optional = true
test.optional = true

[tool.poetry.group.dev.dependencies]
coverage-badge = "^1.1.0"
pre-commit = "^3.3.3"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"


[tool.pytest.ini_options]
addopts = "--cov=grid_hooks --cov-report=term-missing --cov-fail-under=70"
testpaths = ["tests"]


[tool.isort]
profile = "black"
